Optimized netmhcpan for distributed computing
Previously netmhcpan was run on both MUT and wild type in series.  Now with the addition of a channel for inputType which should be 'MUT' or 'WT', it will kick off two jobs.  Just make a Channel.Of('MUT','WT') outside the module as an input and it should work.  The original function of running them in series is kicked off but putting in anything other than MUT or WT.