Initial commit
Previously netmhcstabpan was run on both MUT and wild type in series. Now with the addition of a channel for inputType which should be 'MUT' or 'WT', it will kick off two jobs. Just make a Channel.Of('MUT','WT') outside the module as an input and it should work. The original function of running them in series is kicked off by putting in anything other than MUT or WT.

Also, now it doesn't output binding affinities, just stabilities.